Fixup component REQUIRES, add esp-idf-lib as submodule
This allows a build to succeed from a fresh "git clone
--recursive" against esp-idf v4.4.8.

esp-idf-lib (https://github.com/UncleRus/esp-idf-lib) is required by the
build but not included, hence the submodule.

The relative and precise paths are no longer needed if properly
declaring REQUIRES in CMakeLists.txt for each component. There's a few
(existing) circular dependencies but the build succeeds.

Also add RoomSense_Logo.png, which was missing and caused the build to
fail.