More consistently put together SoftAP SSIDs
Depending on the length of the location tag, the SSID sometimes
contained a literal '\0' or was missing one of the 4 MAC address
nibbles.

In preparation for addressing corruption with longish location tags,
refactor this a bit so that the length will never go beyond 32 bytes,
the max length for an SSID. If the location is too long, it will be
truncated to fit.

Also, add a taskYIELD() to the retry of esp_wifi_set_mode() per 54138b8.
If the higher priority task in the wifi stack needs to do additional
work to move past ESP_ERR_WIFI_STOP_STATE, this should trigger a context
switch to let that happen.