Allow location tags up to 32 characters
Refactor both the HTTP/JSON and NVS handling of the location tag, which
would become corrupt when over 15 characters per
https://github.com/roomsense/firmware/issues/15 due to a lack of NULL
terminating byte causing adjacent uninitialized data to be read.

Since nvs_set_blob / nvs_store_blob is used, treat the 32 byte (prev 16
byte) blob as just that, and ensure that the NULL byte is in the
destination strings before copying.

This also does a bit more error checking in terms of ensuring the
provided JSON variable is indeed a string, is no longer than 32
characters, and also now allows zero-length location tags.

A reboot is still needed to repopulate the MQTT autodiscovery fields and
change the SoftAP SSID.

{0} rather than {} is now used when allocating the roomsense_iq_shared
struct to explicitly state that all bytes should be NULL to begin with;
when updating the tag, it's rezeroed with memset.