Don't submit location data unless in "submit" mode
One side effect of allowing empty locations is that simply clicking the
"Location Tag" button would make a request with empty data that cleared
the existing tag.

Dynamically update the onclick value to distinguish between modes, and
don't submit changes unless the location value itself has changed.

If an error is returned, don't just silently log it to JS console; use
an old school alert() to show it.

Finally, remove the default value of "Not Set" which would flash briefly
when loading the page. If the location is blank, it'll populate after
load.

The null checks in the JSON don't really make sense as a missing value
in NVS leads to an empty string due to how it's put together, but
(slightly cleaned up) conditionals in the JS will handle them.